requirements.txt
colorlog
discord.py
python-dotenv
mcrcon